refactor(bin-designer): share the param-modified compare between the two dot mappings
groupModified and categoryModified held byte-identical canonical/matchesDefault
helpers; extract one paramFieldModified predicate they both call.
